--- a/Cargo.toml
+++ b/Cargo.toml
@@ -25,6 +25,7 @@ arrayvec = "0.5.2"
 [dev-dependencies]
 bincode = "1.3"
 bytes = "1.0"
+csv = "1"
 futures = "0.3"
 rand = "0.7"
 serde_derive = "1.0"