--- a/Cargo.toml
+++ b/Cargo.toml
@@ -36,7 +36,7 @@ tokio = { features = ["rt-multi-thread", "test-util", "macros"], version = "1.0"
 db-diesel-postgres = ["diesel", "std"]
 db-postgres = ["byteorder", "bytes", "postgres", "std"]
 db-tokio-postgres = ["byteorder", "bytes", "postgres", "std", "tokio-postgres"]
-default = ["serde", "std"]
+default = ["fast-div", "serde", "std"]
 fast-div = []
 serde-bincode = ["serde-str"] # Backwards compatability
 serde-float = ["serde"]