This is the start of cleaning up some of the internals to optimize the performance of some of the slower parts of Rust Decimal. This first project attacks division, however there are definitely further tasks that can be optimized including `rem`, `to_string` etc. 

Since it is a huge rewrite of how division works I've added a ton more tests. I've added 70,000 randomized tests using various configurations of `Decimal` to try to capture and curb against any bugs that may creep in. The implementation DOES generate different results than the legacy version in some cases. I've determined that these are most likely rounding issues during overflow/underflow. In these cases the tests are separated out by a new macro `fast_div_fix!`.

Last bu not least: I have made this new implementation the default however you _can_ opt out by using `no-default-features` and leaving out the `fast-div` feature.

Ultimately this resolves #276 and also makes progress against #148.

Before:
```
test add_10k                 ... bench:      40,956 ns/iter (+/- 2,057)
test add_negative_pi         ... bench:         138 ns/iter (+/- 6)
test add_negative_point_five ... bench:          15 ns/iter (+/- 0)
test add_one                 ... bench:          19 ns/iter (+/- 1)
test add_one_hundred         ... bench:          19 ns/iter (+/- 0)
test add_pi                  ... bench:         138 ns/iter (+/- 6)
test add_point_zero_one      ... bench:           7 ns/iter (+/- 0)
test add_two                 ... bench:          19 ns/iter (+/- 0)
test decimal_from_str        ... bench:         478 ns/iter (+/- 26)
test decimal_to_string       ... bench:       2,171 ns/iter (+/- 233)
test div_10k                 ... bench:     192,245 ns/iter (+/- 6,577)
test div_negative_pi         ... bench:      24,904 ns/iter (+/- 1,765)
test div_negative_point_five ... bench:         381 ns/iter (+/- 17)
test div_no_underflow        ... bench:       9,760 ns/iter (+/- 617)
test div_one                 ... bench:         207 ns/iter (+/- 24)
test div_one_hundred         ... bench:         518 ns/iter (+/- 19)
test div_pi                  ... bench:      24,899 ns/iter (+/- 678)
test div_point_zero_one      ... bench:         206 ns/iter (+/- 7)
test div_two                 ... bench:         386 ns/iter (+/- 24)
test erf                     ... bench:     137,342 ns/iter (+/- 6,063)
test exp                     ... bench:     458,736 ns/iter (+/- 12,021)
test iterator_individual     ... bench:          50 ns/iter (+/- 2)
test iterator_sum            ... bench:          50 ns/iter (+/- 1)
test ln                      ... bench:   5,339,187 ns/iter (+/- 178,306)
test mul_negative_pi         ... bench:         154 ns/iter (+/- 21)
test mul_negative_point_five ... bench:           9 ns/iter (+/- 0)
test mul_one                 ... bench:           9 ns/iter (+/- 0)
test mul_one_hundred         ... bench:           9 ns/iter (+/- 0)
test mul_pi                  ... bench:         152 ns/iter (+/- 12)
test mul_point_zero_one      ... bench:           9 ns/iter (+/- 0)
test mul_two                 ... bench:           9 ns/iter (+/- 0)
test norm_cdf                ... bench:     234,461 ns/iter (+/- 12,060)
test norm_pdf                ... bench:     388,250 ns/iter (+/- 16,999)
test powi                    ... bench:         861 ns/iter (+/- 76)
test serialize_bincode       ... bench:       2,147 ns/iter (+/- 123)
test sqrt                    ... bench:   1,067,460 ns/iter (+/- 53,104)
test sub_10k                 ... bench:      47,207 ns/iter (+/- 1,774)
test sub_negative_pi         ... bench:         140 ns/iter (+/- 26)
test sub_negative_point_five ... bench:          14 ns/iter (+/- 0)
test sub_one                 ... bench:          20 ns/iter (+/- 1)
test sub_one_hundred         ... bench:          20 ns/iter (+/- 0)
test sub_pi                  ... bench:         138 ns/iter (+/- 4)
test sub_point_zero_one      ... bench:           8 ns/iter (+/- 0)
test sub_two                 ... bench:          20 ns/iter (+/- 1)
```

After:
```
test add_10k                 ... bench:      40,995 ns/iter (+/- 2,273)
test add_negative_pi         ... bench:         137 ns/iter (+/- 5)
test add_negative_point_five ... bench:          15 ns/iter (+/- 0)
test add_one                 ... bench:          19 ns/iter (+/- 0)
test add_one_hundred         ... bench:          19 ns/iter (+/- 0)
test add_pi                  ... bench:         138 ns/iter (+/- 5)
test add_point_zero_one      ... bench:           7 ns/iter (+/- 0)
test add_two                 ... bench:          19 ns/iter (+/- 0)
test decimal_from_str        ... bench:         478 ns/iter (+/- 40)
test decimal_to_string       ... bench:       2,203 ns/iter (+/- 255)
test div_10k                 ... bench:      85,913 ns/iter (+/- 2,663)
test div_negative_pi         ... bench:          98 ns/iter (+/- 4)
test div_negative_point_five ... bench:          39 ns/iter (+/- 2)
test div_no_underflow        ... bench:          88 ns/iter (+/- 5)
test div_one                 ... bench:          17 ns/iter (+/- 0)
test div_one_hundred         ... bench:          39 ns/iter (+/- 2)
test div_pi                  ... bench:          97 ns/iter (+/- 8)
test div_point_zero_one      ... bench:          17 ns/iter (+/- 0)
test div_two                 ... bench:          39 ns/iter (+/- 1)
test erf                     ... bench:      15,646 ns/iter (+/- 1,065)
test exp                     ... bench:      20,058 ns/iter (+/- 1,417)
test iterator_individual     ... bench:          49 ns/iter (+/- 1)
test iterator_sum            ... bench:          50 ns/iter (+/- 1)
test ln                      ... bench:      68,904 ns/iter (+/- 3,811)
test mul_negative_pi         ... bench:         151 ns/iter (+/- 5)
test mul_negative_point_five ... bench:           9 ns/iter (+/- 0)
test mul_one                 ... bench:           9 ns/iter (+/- 0)
test mul_one_hundred         ... bench:           9 ns/iter (+/- 0)
test mul_pi                  ... bench:         152 ns/iter (+/- 6)
test mul_point_zero_one      ... bench:           9 ns/iter (+/- 0)
test mul_two                 ... bench:           9 ns/iter (+/- 0)
test norm_cdf                ... bench:      36,683 ns/iter (+/- 1,552)
test norm_pdf                ... bench:      13,347 ns/iter (+/- 552)
test powi                    ... bench:         859 ns/iter (+/- 49)
test serialize_bincode       ... bench:       2,201 ns/iter (+/- 92)
test sqrt                    ... bench:       8,435 ns/iter (+/- 440)
test sub_10k                 ... bench:      47,443 ns/iter (+/- 1,738)
test sub_negative_pi         ... bench:         138 ns/iter (+/- 8)
test sub_negative_point_five ... bench:          14 ns/iter (+/- 0)
test sub_one                 ... bench:          20 ns/iter (+/- 0)
test sub_one_hundred         ... bench:          20 ns/iter (+/- 0)
test sub_pi                  ... bench:         138 ns/iter (+/- 10)
test sub_point_zero_one      ... bench:           8 ns/iter (+/- 0)
test sub_two                 ... bench:          20 ns/iter (+/- 1)
```